Improve registration and graphing of nested dependencies
To support scenarios where dependencies may have various levels of
dependency relationships, this builds the dependency graph at
registration time rather than at call-time for Manager.adjacencyList().

It does require that the registration function require a new parameter
to indicate the current parent when adding a dependency or descending
into the graph. This is a change in the API signature of the
Manager class, and subsequently a version bump.